+++ crates/libmandoc-rs/shim/mant_mandoc_shim.h
@@ -131,6 +131,7 @@ enum mant_mandoc_author_mode {
 #define MANT_MANDOC_NODE_DELIMITER_OPEN (1U << 7)
 #define MANT_MANDOC_NODE_DELIMITER_CLOSE (1U << 8)
 #define MANT_MANDOC_NODE_SYNOPSIS_PRETTY (1U << 9)
+#define MANT_MANDOC_NODE_TABLE_START (1U << 10)
 
 struct mant_mandoc_document *mant_mandoc_parse_file(
     const char *, const char *, int, int, const char *);